requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1
rank-bm25==0.2.2
sentence-transformers==2.7.0
faiss-cpu==1.8.0
torch==2.3.1
jsonlines==4.0.0
tqdm==4.66.4
aiosqlite==0.20.0
sqlite-utils==3.36
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
pytest==8.2.0
httpx==0.27.0
python-dotenv==1.0.1
slowapi==0.1.9
wikipedia==1.4.0
